Extracting CashFlow functions from Financial
