+++ src/PhpSpreadsheet/Calculation/Financial.php
@@ -874,6 +874,8 @@ class Financial
      * Excel Function:
      *        IRR(values[,guess])
      *
+     * @Deprecated 1.18.0
+     *
      * @param mixed $values An array or a reference to cells that contain numbers for which you want
      *                                    to calculate the internal rate of return.
      *                                Values must contain at least one positive value and one negative value to
@@ -881,56 +883,13 @@ class Financial
      * @param mixed $guess A number that you guess is close to the result of IRR
      *
      * @return float|string
+     *
+     *@see Financial\CashFlow\Variable\Periodic::rate()
+     *      Use the IRR() method in the Financial\CashFlow\Variable\Periodic class instead
      */
     public static function IRR($values, $guess = 0.1)
     {
-        if (!is_array($values)) {
-            return Functions::VALUE();
-        }
-        $values = Functions::flattenArray($values);
-        $guess = Functions::flattenSingleValue($guess);
-
-        // create an initial range, with a root somewhere between 0 and guess
-        $x1 = 0.0;
-        $x2 = $guess;
-        $f1 = self::NPV($x1, $values);
-        $f2 = self::NPV($x2, $values);
-        for ($i = 0; $i < self::FINANCIAL_MAX_ITERATIONS; ++$i) {
-            if (($f1 * $f2) < 0.0) {
-                break;
-            }
-            if (abs($f1) < abs($f2)) {
-                $f1 = self::NPV($x1 += 1.6 * ($x1 - $x2), $values);
-            } else {
-                $f2 = self::NPV($x2 += 1.6 * ($x2 - $x1), $values);
-            }
-        }
-        if (($f1 * $f2) > 0.0) {
-            return Functions::VALUE();
-        }
-
-        $f = self::NPV($x1, $values);
-        if ($f < 0.0) {
-            $rtb = $x1;
-            $dx = $x2 - $x1;
-        } else {
-            $rtb = $x2;
-            $dx = $x1 - $x2;
-        }
-
-        for ($i = 0; $i < self::FINANCIAL_MAX_ITERATIONS; ++$i) {
-            $dx *= 0.5;
-            $x_mid = $rtb + $dx;
-            $f_mid = self::NPV($x_mid, $values);
-            if ($f_mid <= 0.0) {
-                $rtb = $x_mid;
-            }
-            if ((abs($f_mid) < self::FINANCIAL_PRECISION) || (abs($dx) < self::FINANCIAL_PRECISION)) {
-                return $x_mid;
-            }
-        }
-
-        return Functions::VALUE();
+        return Financial\CashFlow\Variable\Periodic::rate($values, $guess);
     }
 
     /**
@@ -984,6 +943,11 @@ class Financial
      * Excel Function:
      *        MIRR(values,finance_rate, reinvestment_rate)
      *
+     * @Deprecated 1.18.0
+     *
+     * @see Financial\CashFlow\Variable\Periodic::modifiedRate()
+     *      Use the MIRR() method in the Financial\CashFlow\Variable\Periodic class instead
+     *
      * @param mixed $values An array or a reference to cells that contain a series of payments and
      *                         income occurring at regular intervals.
      *                      Payments are negative value, income is positive values.
@@ -994,34 +958,7 @@ class Financial
      */
     public static function MIRR($values, $finance_rate, $reinvestment_rate)
     {
-        if (!is_array($values)) {
-            return Functions::VALUE();
-        }
-        $values = Functions::flattenArray($values);
-        $finance_rate = Functions::flattenSingleValue($finance_rate);
-        $reinvestment_rate = Functions::flattenSingleValue($reinvestment_rate);
-        $n = count($values);
-
-        $rr = 1.0 + $reinvestment_rate;
-        $fr = 1.0 + $finance_rate;
-
-        $npv_pos = $npv_neg = 0.0;
-        foreach ($values as $i => $v) {
-            if ($v >= 0) {
-                $npv_pos += $v / $rr ** $i;
-            } else {
-                $npv_neg += $v / $fr ** $i;
-            }
-        }
-
-        if (($npv_neg == 0) || ($npv_pos == 0) || ($reinvestment_rate <= -1)) {
-            return Functions::VALUE();
-        }
-
-        $mirr = ((-$npv_pos * $rr ** $n)
-                / ($npv_neg * ($rr))) ** (1.0 / ($n - 1)) - 1.0;
-
-        return is_finite($mirr) ? $mirr : Functions::VALUE();
+        return Financial\CashFlow\Variable\Periodic::modifiedRate($values, $finance_rate, $reinvestment_rate);
     }
 
     /**
@@ -1092,28 +1029,16 @@ class Financial
      *
      * Returns the Net Present Value of a cash flow series given a discount rate.
      *
+     * @Deprecated 1.18.0
+     *
      * @return float
+     *
+     *@see Financial\CashFlow\Variable\Periodic::presentValue()
+     *      Use the NPV() method in the Financial\CashFlow\Variable\Periodic class instead
      */
     public static function NPV(...$args)
     {
-        // Return value
-        $returnValue = 0;
-
-        // Loop through arguments
-        $aArgs = Functions::flattenArray($args);
-
-        // Calculate
-        $rate = array_shift($aArgs);
-        $countArgs = count($aArgs);
-        for ($i = 1; $i <= $countArgs; ++$i) {
-            // Is it a numeric value?
-            if (is_numeric($aArgs[$i - 1])) {
-                $returnValue += $aArgs[$i - 1] / (1 + $rate) ** $i;
-            }
-        }
-
-        // Return
-        return $returnValue;
+        return Financial\CashFlow\Variable\Periodic::presentValue(...$args);
     }
 
     /**
@@ -1587,85 +1512,6 @@ class Financial
         return TreasuryBill::yield($settlement, $maturity, $price);
     }
 
-    private static function bothNegAndPos($neg, $pos)
-    {
-        return $neg && $pos;
-    }
-
-    private static function xirrPart2(&$values)
-    {
-        $valCount = count($values);
-        $foundpos = false;
-        $foundneg = false;
-        for ($i = 0; $i < $valCount; ++$i) {
-            $fld = $values[$i];
-            if (!is_numeric($fld)) {
-                return Functions::VALUE();
-            } elseif ($fld > 0) {
-                $foundpos = true;
-            } elseif ($fld < 0) {
-                $foundneg = true;
-            }
-        }
-        if (!self::bothNegAndPos($foundneg, $foundpos)) {
-            return Functions::NAN();
-        }
-
-        return '';
-    }
-
-    private static function xirrPart1(&$values, &$dates)
-    {
-        if ((!is_array($values)) && (!is_array($dates))) {
-            return Functions::NA();
-        }
-        $values = Functions::flattenArray($values);
-        $dates = Functions::flattenArray($dates);
-        if (count($values) != count($dates)) {
-            return Functions::NAN();
-        }
-
-        $datesCount = count($dates);
-        for ($i = 0; $i < $datesCount; ++$i) {
-            try {
-                $dates[$i] = DateTimeExcel\Helpers::getDateValue($dates[$i]);
-            } catch (Exception $e) {
-                return $e->getMessage();
-            }
-        }
-
-        return self::xirrPart2($values);
-    }
-
-    private static function xirrPart3($values, $dates, $x1, $x2)
-    {
-        $f = self::xnpvOrdered($x1, $values, $dates, false);
-        if ($f < 0.0) {
-            $rtb = $x1;
-            $dx = $x2 - $x1;
-        } else {
-            $rtb = $x2;
-            $dx = $x1 - $x2;
-        }
-
-        $rslt = Functions::VALUE();
-        for ($i = 0; $i < self::FINANCIAL_MAX_ITERATIONS; ++$i) {
-            $dx *= 0.5;
-            $x_mid = $rtb + $dx;
-            $f_mid = self::xnpvOrdered($x_mid, $values, $dates, false);
-            if ($f_mid <= 0.0) {
-                $rtb = $x_mid;
-            }
-            if ((abs($f_mid) < self::FINANCIAL_PRECISION) || (abs($dx) < self::FINANCIAL_PRECISION)) {
-                $rslt = $x_mid;
-
-                break;
-            }
-        }
-
-        return $rslt;
-    }
-
     /**
      * XIRR.
      *
@@ -1673,6 +1519,11 @@ class Financial
      *
      * Excel Function:
      *        =XIRR(values,dates,guess)
+     *
+     * @Deprecated 1.18.0
+     *
+     * @see Financial\CashFlow\Variable\NonPeriodic::rate()
+     *      Use the rate() method in the Financial\CashFlow\Variable\NonPeriodic class instead
      *
      * @param float[] $values     A series of cash flow payments
      *                                The series of values must contain at least one positive value & one negative value
@@ -1685,37 +1536,7 @@ class Financial
      */
     public static function XIRR($values, $dates, $guess = 0.1)
     {
-        $rslt = self::xirrPart1($values, $dates);
-        if ($rslt) {
-            return $rslt;
-        }
-
-        // create an initial range, with a root somewhere between 0 and guess
-        $guess = Functions::flattenSingleValue($guess);
-        $x1 = 0.0;
-        $x2 = $guess ?: 0.1;
-        $f1 = self::xnpvOrdered($x1, $values, $dates, false);
-        $f2 = self::xnpvOrdered($x2, $values, $dates, false);
-        $found = false;
-        for ($i = 0; $i < self::FINANCIAL_MAX_ITERATIONS; ++$i) {
-            if (!is_numeric($f1) || !is_numeric($f2)) {
-                break;
-            }
-            if (($f1 * $f2) < 0.0) {
-                $found = true;
-
-                break;
-            } elseif (abs($f1) < abs($f2)) {
-                $f1 = self::xnpvOrdered($x1 += 1.6 * ($x1 - $x2), $values, $dates, false);
-            } else {
-                $f2 = self::xnpvOrdered($x2 += 1.6 * ($x2 - $x1), $values, $dates, false);
-            }
-        }
-        if (!$found) {
-            return Functions::NAN();
-        }
-
-        return self::xirrPart3($values, $dates, $x1, $x2);
+        return Financial\CashFlow\Variable\NonPeriodic::rate($values, $dates, $guess);
     }
 
     /**
@@ -1726,6 +1547,11 @@ class Financial
      *
      * Excel Function:
      *        =XNPV(rate,values,dates)
+     *
+     * @Deprecated 1.18.0
+     *
+     * @see Financial\CashFlow\Variable\NonPeriodic::presentValue()
+     *      Use the presentValue() method in the Financial\CashFlow\Variable\NonPeriodic class instead
      *
      * @param float $rate the discount rate to apply to the cash flows
      * @param float[] $values     A series of cash flows that corresponds to a schedule of payments in dates.
@@ -1740,68 +1566,7 @@ class Financial
      */
     public static function XNPV($rate, $values, $dates)
     {
-        return self::xnpvOrdered($rate, $values, $dates, true);
-    }
-
-    private static function validateXnpv($rate, $values, $dates)
-    {
-        if (!is_numeric($rate)) {
-            return Functions::VALUE();
-        }
-        $valCount = count($values);
-        if ($valCount != count($dates)) {
-            return Functions::NAN();
-        }
-        if ($valCount > 1 && ((min($values) > 0) || (max($values) < 0))) {
-            return Functions::NAN();
-        }
-        $date0 = DateTimeExcel\Helpers::getDateValue($dates[0]);
-        if (is_string($date0)) {
-            return Functions::VALUE();
-        }
-
-        return '';
-    }
-
-    private static function xnpvOrdered($rate, $values, $dates, $ordered = true)
-    {
-        $rate = Functions::flattenSingleValue($rate);
-        $values = Functions::flattenArray($values);
-        $dates = Functions::flattenArray($dates);
-        $valCount = count($values);
-
-        try {
-            $date0 = DateTimeExcel\Helpers::getDateValue($dates[0]);
-        } catch (Exception $e) {
-            return $e->getMessage();
-        }
-        $rslt = self::validateXnpv($rate, $values, $dates);
-        if ($rslt) {
-            return $rslt;
-        }
-        $xnpv = 0.0;
-        for ($i = 0; $i < $valCount; ++$i) {
-            if (!is_numeric($values[$i])) {
-                return Functions::VALUE();
-            }
-
-            try {
-                $datei = DateTimeExcel\Helpers::getDateValue($dates[$i]);
-            } catch (Exception $e) {
-                return $e->getMessage();
-            }
-            if ($date0 > $datei) {
-                $dif = $ordered ? Functions::NAN() : -DateTimeExcel\DateDif::funcDateDif($datei, $date0, 'd');
-            } else {
-                $dif = DateTimeExcel\DateDif::funcDateDif($date0, $datei, 'd');
-            }
-            if (!is_numeric($dif)) {
-                return $dif;
-            }
-            $xnpv += $values[$i] / (1 + $rate) ** ($dif / 365);
-        }
-
-        return is_finite($xnpv) ? $xnpv : Functions::VALUE();
+        return Financial\CashFlow\Variable\NonPeriodic::presentValue($rate, $values, $dates);
     }
 
     /**
@@ -1809,7 +1574,10 @@ class Financial
      *
      * Returns the annual yield of a security that pays interest at maturity.
      *
-     * @see Use the yieldDiscounted() method in the Financial\Securities\Yields class instead
+     * @Deprecated 1.18.0
+     *
+     * @see Financial\Securities\Yields::yieldDiscounted()
+     *      Use the yieldDiscounted() method in the Financial\Securities\Yields class instead
      *
      * @param mixed $settlement The security's settlement date.
      *                              The security's settlement date is the date after the issue date when the security
@@ -1839,7 +1607,8 @@ class Financial
      *
      * @Deprecated 1.18.0
      *
-     * @see Use the yieldAtMaturity() method in the Financial\Securities\Yields class instead
+     * @see Financial\Securities\Yields::yieldAtMaturity()
+     *      Use the yieldAtMaturity() method in the Financial\Securities\Yields class instead
      *
      * @param mixed $settlement The security's settlement date.
      *                              The security's settlement date is the date after the issue date when the security
