+++ tests/data/Calculation/Financial/FVSCHEDULE.php
@@ -36,4 +36,22 @@ return [
             ],
         ],
     ],
+    [
+        '#VALUE!',
+        'NaN',
+        [
+            0.089999999999999997,
+            0.11,
+            0.10000000000000001,
+        ],
+    ],
+    [
+        '#VALUE!',
+        100,
+        [
+            0.089999999999999997,
+            'NaN',
+            0.10000000000000001,
+        ],
+    ],
 ];
