+++ tests/data/Calculation/Financial/IRR.php
@@ -32,7 +32,7 @@ return [
             15000,
             18000,
         ],
-        0.10000000000000001,
+        0.10,
     ],
     [
         -0.13618951095869,
@@ -41,7 +41,7 @@ return [
             [
                 20,
                 24,
-                28.800000000000001,
+                28.8,
             ],
         ],
     ],
@@ -52,9 +52,9 @@ return [
             [
                 20,
                 24,
-                28.800000000000001,
-                34.560000000000002,
-                41.469999999999999,
+                28.8,
+                34.56,
+                41.47,
             ],
         ],
     ],
