+++ tests/data/Calculation/Financial/MIRR.php
@@ -15,7 +15,7 @@ return [
                 46000,
             ],
         ],
-        0.10000000000000001,
+        0.10,
         0.12,
     ],
     [
@@ -28,7 +28,7 @@ return [
                 21000,
             ],
         ],
-        0.10000000000000001,
+        0.10,
         0.12,
     ],
     [
@@ -43,8 +43,8 @@ return [
                 46000,
             ],
         ],
-        0.10000000000000001,
-        0.14000000000000001,
+        0.10,
+        0.14,
     ],
     [
         0.74021752686287001,
