+++ tests/data/Calculation/Financial/NPV.php
@@ -5,7 +5,7 @@
 return [
     [
         1188.4434123352,
-        0.10000000000000001,
+        0.10,
         -10000,
         3000,
         4200,
@@ -13,7 +13,7 @@ return [
     ],
     [
         41922.061554931999,
-        0.080000000000000002,
+        0.08,
         8000,
         9200,
         10000,
@@ -22,7 +22,7 @@ return [
     ],
     [
         36250.534912984003,
-        0.080000000000000002,
+        0.08,
         8000,
         9200,
         10000,
@@ -32,7 +32,7 @@ return [
     ],
     [
         12678.677633095,
-        0.050000000000000003,
+        0.05,
         2000,
         2400,
         2900,
