+++ tests/data/Calculation/Financial/PDURATION.php
@@ -2,18 +2,6 @@
 
 return [
     [
-        '#VALUE!',
-        ['ABC', 'DEF', 'GHI'],
-    ],
-    [
-        '#NUM!',
-        [0.0, 0.0, 0.0],
-    ],
-    [
-        '#NUM!',
-        [0.05, 10.0, -2.0],
-    ],
-    [
         10.33803507,
         [0.04, 10000, 15000],
     ],
@@ -32,5 +20,29 @@ return [
     [
         14.206699082890474,
         [0.05, 50, 100],
+    ],
+    [
+        '#VALUE!',
+        ['NaN', 50, 100],
+    ],
+    [
+        '#VALUE!',
+        [0.05, 'NaN', 100],
+    ],
+    [
+        '#VALUE!',
+        [0.05, 50, 'NaN'],
+    ],
+    [
+        '#NUM!',
+        [0.00, 50, 100],
+    ],
+    [
+        '#NUM!',
+        [0.05, 0, 100],
+    ],
+    [
+        '#NUM!',
+        [0.05, 50, 0],
     ],
 ];
