+++ tests/data/Calculation/Financial/RRI.php
@@ -2,18 +2,6 @@
 
 return [
     [
-        '#VALUE!',
-        ['ABC', 'DEF', 'GHI'],
-    ],
-    [
-        '#NUM!',
-        [0.0, 10.0, 20.0],
-    ],
-    [
-        '#NUM!',
-        [0.05, 10.0, -2.0],
-    ],
-    [
         0.04137974399241062,
         [10, 10000, 15000],
     ],
@@ -24,5 +12,29 @@ return [
     [
         0.6747967875721199,
         [10, 70, 12154],
+    ],
+    [
+        '#VALUE!',
+        ['NaN', 70, 12154],
+    ],
+    [
+        '#VALUE!',
+        [10, 'NaN', 12154],
+    ],
+    [
+        '#VALUE!',
+        [10, 70, 'NaN'],
+    ],
+    [
+        '#NUM!',
+        [0.0, 70.0, 200.0],
+    ],
+    [
+        '#NUM!',
+        [10, 0.0, 200.0],
+    ],
+    [
+        '#NUM!',
+        [10, 70, -2.0],
     ],
 ];
